Fix version issues in a clean environment
We were attempting to do a solution described here, and copied from
keras-tuner and autokeras:
https://packaging.python.org/en/latest/guides/single-sourcing-package-version/

But the approach relies on a certain version of setuptools beyond what
is default on most default OS python installs.

Simplest approach seems to be to declare the version in two places, so
let's just do that for now.