Return a single tensor from roberta (#328)
I think as an overall policy, it makes sense to return a single tensor
as a tensor not as a dictionary.

We could consider a dict to allow for future modifications to the output
to support more types. But I feel like that would be incorrect in this
case. Different "heads" for this base graph might define different
outputs, and this graph could be sliced to grab intermediate outputs,
but there will only ever be a single output from the RoBERTa core graph.