Add a test harness based on keras-core's `run_layer_test` (#1238)
* Add a test harness based on keras-core's `run_layer_test`

Eventually, I want to add some dtype changes similar to ->
https://github.com/keras-team/keras-core/pull/805
But the nice for that PR on keras-core was I could add dtype test to a
common harness and test all layers.

So I think it's finally time to bite the bullet and add these for
keras-nlp. This ports and simplifies the `run_layer_test` code from
keras-core and applies it to our modeling layers.

I am ditching the saved model tests for our individual layers, with
the idea being that saved model tests are slow, and we get fairly
robust serialization tests now without saved model. If this is good
enough for keras-core layers, I think we can follow suit here. We still
test saving end to end through our modeling tests.

* Fix tests

* Address comments