Remove the score function from pali gemma causal lm
I am not sure if we want to force the users to pass the image
embeddings themselves, it might be more frienly to allow the raw
image input.

Anyway, since this is unused and untested, and we should check with
the LIT team anyway on how to set this up, let's just remove it for
now.